An a private function called "addNickname(view: View)" was added the MainActivity.kt
This function set our nickName text View equal to the value of the edit View that the user entered. Furthermore, the editView, and the "done" button View were made invisible, while the "Nickname" text View was made visible

An onClickListener was added to the MainActivity.kt at the end of the onCreate function. It served the purpose of referencing our button View and taking it in as a parameter for our "addNickName" function